Review fixes Rename package, validate config, code style.
