+++ src/test/kotlin/dsx/bps/core/InvoiceProcessorUnitTest.kt
@@ -3,15 +3,13 @@ package dsx.bps.core
 import com.uchuhimo.konf.Config
 import com.uchuhimo.konf.source.yaml
 import dsx.bps.config.InvoiceProcessorConfig
-import dsx.bps.config.currencyconfig.TrxConfig
 import dsx.bps.core.datamodel.*
-import dsx.bps.crypto.trx.TrxClient
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
-import org.mockito.Mockito;
+import org.mockito.Mockito
 import io.reactivex.disposables.Disposable
 import org.junit.jupiter.api.Nested
 import java.io.File
@@ -35,8 +33,6 @@ internal class InvoiceProcessorUnitTest {
 
         invoiceProcessor = InvoiceProcessor(manager, testConfig)
     }
-
-
 
     @Test
     @DisplayName("create invoice and get invoice test")
